Penalized Functional Dirichlet Process Linear Mixed Effects Regression with Between-Within Decomposition

fits a functional dirichlet process linear mixed effects regression model with N observations and n subjects using a between-within effects decomposition on the STAP-DP design matrix

stappDP_merdecomp(
  y,
  Z,
  X_b,
  X_w,
  W,
  S_b,
  S_w,
  w,
  subj_mat_,
  subj_n,
  alpha_a,
  alpha_b,
  sigma_a,
  sigma_b,
  tau_a,
  tau_b,
  K,
  num_penalties,
  iter_max,
  burn_in,
  thin,
  seed,
  num_posterior_samples,
  fix_alpha
)

Arguments

y
a vector of continuous outcomes

Z
a matrix of population level confounders

X_b
Matrix of between subject spatial temporal aggregated predictor covariates

X_w
Matrix of within subject spatial temporal aggregated predictor covariates

W
a design matrix for group specific terms

S_b
penalty matrix corresponding to between subject covariate matrix

S_w
penalty matrix corresponding to within subject covariate matrix

w
a vector of weights for weighted regression

subj_mat_
N x n sparse matrix used to aggregate subject observations

subj_n
n x 1 vector of integers representing how many observations correspond to each subject

alpha_a
alpha gamma prior shape hyperparameter

alpha_b
alpha gamma prior scale hyperparameter

sigma_a
precision gamma prior shape hyperparameter

sigma_b
precision gamma prior scale hyperparameter

tau_a
penalty gamma prior shape hyperparameter

tau_b
penalty gamma prior scale hyperparameter

K
truncation number

num_penalties
number of penalty matrices accounted for in S

iter_max
maximum number of iterations

burn_in
number of burn in iterations

thin
number by which to thin samples

seed
rng initializer

num_posterior_samples
total number of posterior samples

fix_alpha
boolean value that determines whether or not to fix alpha in sampler